requirements.txt
discord.py>=2.3.2
twitchio>=2.8.2
flask>=2.0.1
requests>=2.31.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
gunicorn>=20.1.0
